requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
parsel>=1.9
anthropic>=0.40
deep-translator>=1.11
python-telegram-bot>=21.0
python-dotenv>=1.0
pydantic>=2.0
loguru>=0.7
